+++ cw-vault-standard/src/msg.rs
@@ -90,44 +90,6 @@ where
     #[returns(VaultInfoResponse)]
     Info {},
 
-    /// Returns `Uint128` amount of vault tokens that will be returned for the
-    /// passed in `amount` of base tokens.
-    ///
-    /// Allows an on-chain or off-chain user to simulate the effects of their
-    /// deposit at the current block, given current on-chain conditions.
-    ///
-    /// Must return as close to and no more than the exact amount of vault
-    /// tokens that would be minted in a deposit call in the same transaction.
-    /// I.e. Deposit should return the same or more vault tokens as
-    /// PreviewDeposit if called in the same transaction.
-    #[deprecated(
-        since = "0.4.1",
-        note = "PreviewDeposit and PreviewRedeem turned out to be too difficult to implement in most cases. We recommend to use transaction simulation from non-contract clients such as frontends."
-    )]
-    #[returns(Uint128)]
-    PreviewDeposit {
-        /// The amount of base tokens to preview depositing.
-        amount: Uint128,
-    },
-
-    /// Returns `Uint128` amount of base tokens that would be withdrawn in
-    /// exchange for redeeming `amount` of vault tokens.
-    ///
-    /// Allows an on-chain or off-chain user to simulate the effects of their
-    /// redeem at the current block, given current on-chain conditions.
-    ///
-    /// Must return as close to and no more than the exact amount of base tokens
-    /// that would be withdrawn in a redeem call in the same transaction.
-    #[deprecated(
-        since = "0.4.1",
-        note = "PreviewDeposit and PreviewRedeem turned out to be too difficult to implement in most cases. We recommend to use transaction simulation from non-contract clients such as frontends."
-    )]
-    #[returns(Uint128)]
-    PreviewRedeem {
-        /// The amount of vault tokens to preview redeeming.
-        amount: Uint128,
-    },
-
     /// Returns the amount of assets managed by the vault denominated in base
     /// tokens. Useful for display purposes, and does not have to confer the
     /// exact amount of base tokens.